oops

+++ arrow-data-source/common/src/main/scala/org/apache/spark/sql/execution/datasources/v2/arrow/SparkMemoryUtils.scala
@@ -46,7 +46,7 @@ object SparkMemoryUtils extends Logging {
       val al = new SparkManagedAllocationListener(
         new NativeSQLMemoryConsumer(getTaskMemoryManager(), Spiller.NO_OP),
         sharedMetrics)
-      new RootAllocator(al, Int.MaxValue)
+      new RootAllocator(al, Long.MaxValue)
     }
 
     val defaultMemoryPool: NativeMemoryPoolWrapper = {
